@@ -3,10 +3,11 @@
 buildscript {
     repositories {
         jcenter()
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.1'
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:1.5'
+        classpath 'com.android.tools.build:gradle:3.2.0'
+        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.0'
     }
 }
 
@@ -14,15 +15,20 @@ allprojects {
     repositories {
         jcenter()
         maven { url "https://jitpack.io" }
+        google()
     }
 }
 
 ext {
-    compileSdkVersion = 25
-    buildToolsVersion = '25.0.3'
+    compileSdkVersion = 28
+    buildToolsVersion = '28.0.0'
     targetSdkVersion = compileSdkVersion
 
-    supportLibVersion = '25.3.1'
+    appcompatLibVersion = '1.1.0'
+    cardViewLibVersion = '1.0.0'
+    materialLibVersion = '1.0.0'
+    iconHandlerVersion = 'v2.0.0'
+    butterKnifeVersion = '8.8.1'
 }
 
 task clean(type: Delete) {
