@@ -5,27 +5,31 @@ import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
-import android.support.v4.content.ContextCompat;
+import androidx.annotation.DrawableRes;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.core.content.ContextCompat;
 
 /**
- * Created by jrvansuita on 10/02/17.
+ * Static utilities.
  */
-
+@SuppressWarnings("unused")
 public final class IconUtil {
 
     public static Bitmap getBitmap(Context context, int res) {
         return BitmapFactory.decodeResource(context.getResources(), res);
     }
 
-    public static Bitmap getBitmap(Context context, Drawable drawable) {
-        return ((BitmapDrawable) drawable).getBitmap();
+    public static Bitmap getBitmap(BitmapDrawable drawable) {
+        return drawable.getBitmap();
     }
 
     public static Drawable getDrawable(Context context, Bitmap bitmap) {
         return new BitmapDrawable(context.getResources(), bitmap);
     }
 
-    public static Drawable getDrawable(Context context, int res) {
+    @Nullable
+    public static Drawable getDrawable(@NonNull Context context, @DrawableRes int res) {
         return ContextCompat.getDrawable(context, res);
     }
 }
