@@ -1,14 +1,16 @@
 package com.vansuita.materialabout.builder;
 
-
 import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
-import android.graphics.drawable.Drawable;
+import android.graphics.drawable.BitmapDrawable;
 import android.net.Uri;
+import androidx.annotation.NonNull;
 import android.view.View;
+
+import androidx.cardview.widget.CardView;
 
 import com.vansuita.materialabout.R;
 import com.vansuita.materialabout.util.ColorUtil;
@@ -21,9 +23,9 @@ import java.util.LinkedList;
 import static com.vansuita.materialabout.R.mipmap.share;
 
 /**
- * Created by jrvansuita on 10/02/17.
+ * Used to build an {@link AboutView}.
  */
-
+@SuppressWarnings({"WeakerAccess", "unused"})
 public final class AboutBuilder {
 
     private Context context;
@@ -58,15 +60,20 @@ public final class AboutBuilder {
     private boolean wrapScrollView = false;
     private boolean showAsCard = true;
 
-    private LinkedList<Item> links = new LinkedList();
-    private LinkedList<Item> actions = new LinkedList();
+    private LinkedList<Item> links = new LinkedList<>();
+    private LinkedList<Item> actions = new LinkedList<>();
 
+    /**
+     * @deprecated Used {@link #with(Context)} instead.
+     */
+    @Deprecated
     AboutBuilder(Context context) {
         this.context = context;
         this.util = new IntentUtil(context);
     }
 
     public static AboutBuilder with(Context context) {
+        //noinspection deprecation
         return new AboutBuilder(context);
     }
 
@@ -112,6 +119,7 @@ public final class AboutBuilder {
      * @param text the name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setName(String text) {
         this.name = text;
         return this;
@@ -123,10 +131,10 @@ public final class AboutBuilder {
      * @param text the name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setName(int text) {
         return setName(context.getString(text));
     }
-
 
     /**
      * Sets the sub title. It will be place below the developer name
@@ -134,6 +142,7 @@ public final class AboutBuilder {
      * @param text the sub title
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setSubTitle(String text) {
         this.subTitle = text;
         return this;
@@ -145,6 +154,7 @@ public final class AboutBuilder {
      * @param text the sub title
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setSubTitle(int text) {
         return setSubTitle(context.getString(text));
     }
@@ -155,11 +165,11 @@ public final class AboutBuilder {
      * @param text the brief
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setBrief(String text) {
         this.brief = text;
         return this;
     }
-
 
     /**
      * Sets a personal brief
@@ -167,10 +177,10 @@ public final class AboutBuilder {
      * @param text the brief
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setBrief(int text) {
         return setBrief(context.getString(text));
     }
-
 
     /**
      * Sets the app name
@@ -178,6 +188,7 @@ public final class AboutBuilder {
      * @param text the app name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setAppName(String text) {
         this.appName = text;
         return this;
@@ -189,10 +200,10 @@ public final class AboutBuilder {
      * @param text the app name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setAppName(int text) {
         return setAppName(context.getString(text));
     }
-
 
     /**
      * Sets the app title
@@ -200,7 +211,7 @@ public final class AboutBuilder {
      * @param text the title
      * @return the same {@link AboutBuilder} instance
      */
-
+    @NonNull
     public AboutBuilder setAppTitle(String text) {
         this.appTitle = text;
         return this;
@@ -212,6 +223,7 @@ public final class AboutBuilder {
      * @param text the title
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setAppTitle(int text) {
         return setAppTitle(context.getString(text));
     }
@@ -221,6 +233,7 @@ public final class AboutBuilder {
      *
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setVersionNameAsAppSubTitle() {
         try {
             return setAppTitle(context.getString(com.vansuita.materialabout.R.string.version, getPackageInfo().versionName));
@@ -235,6 +248,7 @@ public final class AboutBuilder {
      * @param photo the image
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setPhoto(Bitmap photo) {
         this.photo = photo;
         return this;
@@ -246,6 +260,7 @@ public final class AboutBuilder {
      * @param photo the image
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setPhoto(int photo) {
         return setPhoto(IconUtil.getBitmap(context, photo));
     }
@@ -256,8 +271,9 @@ public final class AboutBuilder {
      * @param photo the image
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder setPhoto(Drawable photo) {
-        return setPhoto(IconUtil.getBitmap(context, photo));
+    @NonNull
+    public AboutBuilder setPhoto(@NonNull BitmapDrawable photo) {
+        return setPhoto(IconUtil.getBitmap(photo));
     }
 
     /**
@@ -266,6 +282,7 @@ public final class AboutBuilder {
      * @param cover the image
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setCover(Bitmap cover) {
         this.cover = cover;
         return this;
@@ -277,6 +294,7 @@ public final class AboutBuilder {
      * @param cover the image
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setCover(int cover) {
         return setCover(IconUtil.getBitmap(context, cover));
     }
@@ -287,8 +305,9 @@ public final class AboutBuilder {
      * @param cover the image
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder setCover(Drawable cover) {
-        return setCover(IconUtil.getBitmap(context, cover));
+    @NonNull
+    public AboutBuilder setCover(@NonNull BitmapDrawable cover) {
+        return setCover(IconUtil.getBitmap(cover));
     }
 
     /**
@@ -297,6 +316,7 @@ public final class AboutBuilder {
      * @param icon the app icon
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setAppIcon(Bitmap icon) {
         this.appIcon = icon;
         return this;
@@ -308,6 +328,7 @@ public final class AboutBuilder {
      * @param icon the app icon
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setAppIcon(int icon) {
         return setAppIcon(IconUtil.getBitmap(context, icon));
     }
@@ -318,8 +339,9 @@ public final class AboutBuilder {
      * @param icon the app icon
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder setAppIcon(Drawable icon) {
-        return setAppIcon(IconUtil.getBitmap(context, icon));
+    @NonNull
+    public AboutBuilder setAppIcon(@NonNull BitmapDrawable icon) {
+        return setAppIcon(IconUtil.getBitmap(icon));
     }
 
     /**
@@ -328,6 +350,7 @@ public final class AboutBuilder {
      * @param color the color resource or the real color.
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setNameColor(int color) {
         this.nameColor = ColorUtil.get(context, color);
         return this;
@@ -339,6 +362,7 @@ public final class AboutBuilder {
      * @param color the color resource or the real color.
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setSubTitleColor(int color) {
         this.subTitleColor = ColorUtil.get(context, color);
         return this;
@@ -350,6 +374,7 @@ public final class AboutBuilder {
      * @param color the color resource or the real color.
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setBriefColor(int color) {
         this.briefColor = ColorUtil.get(context, color);
         return this;
@@ -361,6 +386,7 @@ public final class AboutBuilder {
      * @param color the color resource or the real color.
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setDividerColor(int color) {
         this.dividerColor = ColorUtil.get(context, color);
         return this;
@@ -372,6 +398,7 @@ public final class AboutBuilder {
      * @param color the color resource or the real color.
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setIconColor(int color) {
         this.iconColor = ColorUtil.get(context, color);
         return this;
@@ -383,6 +410,7 @@ public final class AboutBuilder {
      * @param color the color resource or the real color.
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setBackgroundColor(int color) {
         this.backgroundColor = ColorUtil.get(context, color);
         return this;
@@ -394,6 +422,7 @@ public final class AboutBuilder {
      * @param count number of columns
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setActionsColumnsCount(int count) {
         this.actionsColumnsCount = count;
         return this;
@@ -406,6 +435,7 @@ public final class AboutBuilder {
      * @param count number of columns
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setLinksColumnsCount(int count) {
         this.linksColumnsCount = count;
         return this;
@@ -417,6 +447,7 @@ public final class AboutBuilder {
      * @param animate true if you want it
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setLinksAnimated(boolean animate) {
         this.linksAnimated = animate;
         return this;
@@ -428,6 +459,7 @@ public final class AboutBuilder {
      * @param dividerHeight size of the height
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setDividerHeight(int dividerHeight) {
         this.dividerHeight = dividerHeight;
         return this;
@@ -439,6 +471,7 @@ public final class AboutBuilder {
      * @param dividerDashWidth size of the width
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setDividerDashWidth(int dividerDashWidth) {
         this.dividerDashWidth = dividerDashWidth;
         return this;
@@ -450,6 +483,7 @@ public final class AboutBuilder {
      * @param dividerDashGap size of the gap
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setDividerDashGap(int dividerDashGap) {
         this.dividerDashGap = dividerDashGap;
         return this;
@@ -461,6 +495,7 @@ public final class AboutBuilder {
      * @param showDivider true if you want it
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setShowDivider(boolean showDivider) {
         this.showDivider = showDivider;
         return this;
@@ -472,6 +507,7 @@ public final class AboutBuilder {
      * @param wrapScrollView true if you want to wrap
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder setWrapScrollView(boolean wrapScrollView) {
         this.wrapScrollView = wrapScrollView;
         return this;
@@ -485,6 +521,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(Bitmap icon, String label, View.OnClickListener onClickListener) {
         links.add(new Item(icon, label, onClickListener));
         return this;
@@ -498,6 +535,7 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(Bitmap icon, String label, Intent intent) {
         return addLink(icon, label, util.clickIntent(intent));
     }
@@ -510,6 +548,7 @@ public final class AboutBuilder {
      * @param uri   the action uri
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(Bitmap icon, String label, Uri uri) {
         return addLink(icon, label, util.clickUri(uri));
     }
@@ -522,6 +561,7 @@ public final class AboutBuilder {
      * @param url   any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(Bitmap icon, String label, String url) {
         return addLink(icon, label, Uri.parse(url));
     }
@@ -534,6 +574,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(Bitmap icon, int label, View.OnClickListener onClickListener) {
         return addLink(icon, context.getString(label), onClickListener);
     }
@@ -546,6 +587,7 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(Bitmap icon, int label, Intent intent) {
         return addLink(icon, label, util.clickIntent(intent));
     }
@@ -558,6 +600,7 @@ public final class AboutBuilder {
      * @param uri   the action uri
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(Bitmap icon, int label, Uri uri) {
         return addLink(icon, label, util.clickUri(uri));
     }
@@ -570,6 +613,7 @@ public final class AboutBuilder {
      * @param url   any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(Bitmap icon, int label, String url) {
         return addLink(icon, label, Uri.parse(url));
     }
@@ -582,10 +626,10 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(int icon, int label, View.OnClickListener onClickListener) {
         return addLink(IconUtil.getBitmap(context, icon), context.getString(label), onClickListener);
     }
-
 
     /**
      * Adds an link on the links section.
@@ -595,6 +639,7 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(int icon, int label, Intent intent) {
         return addLink(icon, label, util.clickIntent(intent));
     }
@@ -607,6 +652,7 @@ public final class AboutBuilder {
      * @param uri   the action uri
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(int icon, int label, Uri uri) {
         return addLink(icon, label, util.clickUri(uri));
     }
@@ -619,6 +665,7 @@ public final class AboutBuilder {
      * @param url   any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(int icon, int label, String url) {
         return addLink(icon, label, Uri.parse(url));
     }
@@ -631,6 +678,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(int icon, String label, View.OnClickListener onClickListener) {
         return addLink(IconUtil.getBitmap(context, icon), label, onClickListener);
     }
@@ -643,6 +691,7 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(int icon, String label, Intent intent) {
         return addLink(icon, label, util.clickIntent(intent));
     }
@@ -655,6 +704,7 @@ public final class AboutBuilder {
      * @param uri   the action uri
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(int icon, String label, Uri uri) {
         return addLink(icon, label, util.clickUri(uri));
     }
@@ -667,6 +717,7 @@ public final class AboutBuilder {
      * @param url   any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLink(int icon, String label, String url) {
         return addLink(icon, label, Uri.parse(url));
     }
@@ -679,8 +730,9 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addLink(Drawable icon, int label, View.OnClickListener onClickListener) {
-        return addLink(IconUtil.getBitmap(context, icon), context.getString(label), onClickListener);
+    @NonNull
+    public AboutBuilder addLink(@NonNull BitmapDrawable icon, int label, View.OnClickListener onClickListener) {
+        return addLink(IconUtil.getBitmap(icon), context.getString(label), onClickListener);
     }
 
     /**
@@ -691,7 +743,8 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addLink(Drawable icon, int label, Intent intent) {
+    @NonNull
+    public AboutBuilder addLink(@NonNull BitmapDrawable icon, int label, Intent intent) {
         return addLink(icon, label, util.clickIntent(intent));
     }
 
@@ -703,7 +756,8 @@ public final class AboutBuilder {
      * @param uri   the action uri
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addLink(Drawable icon, int label, Uri uri) {
+    @NonNull
+    public AboutBuilder addLink(@NonNull BitmapDrawable icon, int label, Uri uri) {
         return addLink(icon, label, util.clickUri(uri));
     }
 
@@ -715,7 +769,8 @@ public final class AboutBuilder {
      * @param url   any url
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addLink(Drawable icon, int label, String url) {
+    @NonNull
+    public AboutBuilder addLink(@NonNull BitmapDrawable icon, int label, String url) {
         return addLink(icon, label, Uri.parse(url));
     }
 
@@ -727,8 +782,9 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addLink(Drawable icon, String label, View.OnClickListener onClickListener) {
-        return addLink(IconUtil.getBitmap(context, icon), label, onClickListener);
+    @NonNull
+    public AboutBuilder addLink(@NonNull BitmapDrawable icon, String label, View.OnClickListener onClickListener) {
+        return addLink(IconUtil.getBitmap(icon), label, onClickListener);
     }
 
     /**
@@ -739,7 +795,8 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addLink(Drawable icon, String label, Intent intent) {
+    @NonNull
+    public AboutBuilder addLink(@NonNull BitmapDrawable icon, String label, Intent intent) {
         return addLink(icon, label, util.clickIntent(intent));
     }
 
@@ -751,7 +808,8 @@ public final class AboutBuilder {
      * @param uri   the action uri
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addLink(Drawable icon, String label, Uri uri) {
+    @NonNull
+    public AboutBuilder addLink(@NonNull BitmapDrawable icon, String label, Uri uri) {
         return addLink(icon, label, util.clickUri(uri));
     }
 
@@ -763,7 +821,8 @@ public final class AboutBuilder {
      * @param url   any url
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addLink(Drawable icon, String label, String url) {
+    @NonNull
+    public AboutBuilder addLink(@NonNull BitmapDrawable icon, String label, String url) {
         return addLink(icon, label, Uri.parse(url));
     }
 
@@ -774,6 +833,7 @@ public final class AboutBuilder {
      * @param user a GitHub user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addGitHubLink(int user) {
         return addGitHubLink(context.getString(user));
     }
@@ -784,6 +844,7 @@ public final class AboutBuilder {
      * @param user a GitHub user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addGitHubLink(String user) {
         return addLink(R.mipmap.github, R.string.github, util.uri(R.string.url_github, user));
     }
@@ -794,6 +855,7 @@ public final class AboutBuilder {
      * @param user a Bitbucket user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addBitbucketLink(int user) {
         return addBitbucketLink(context.getString(user));
     }
@@ -804,6 +866,7 @@ public final class AboutBuilder {
      * @param user a Bitbucket user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addBitbucketLink(String user) {
         return addLink(R.mipmap.bitbucket, R.string.bitbucket, util.uri(R.string.url_bitbucket, user));
     }
@@ -814,6 +877,7 @@ public final class AboutBuilder {
      * @param user a Facebook user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFacebookLink(int user) {
         return addFacebookLink(context.getString(user));
     }
@@ -824,6 +888,7 @@ public final class AboutBuilder {
      * @param user a Facebook user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFacebookLink(String user) {
         return addLink(R.mipmap.facebook, R.string.facebook, util.openFacebook(user));
     }
@@ -834,6 +899,7 @@ public final class AboutBuilder {
      * @param user a Instagram user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addInstagramLink(int user) {
         return addInstagramLink(context.getString(user));
     }
@@ -844,6 +910,7 @@ public final class AboutBuilder {
      * @param user a Instagram user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addInstagramLink(String user) {
         return addLink(R.mipmap.instagram, R.string.instagram, util.openInstagram(user));
     }
@@ -854,6 +921,7 @@ public final class AboutBuilder {
      * @param user a Twitter user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addTwitterLink(int user) {
         return addTwitterLink(context.getString(user));
     }
@@ -864,6 +932,7 @@ public final class AboutBuilder {
      * @param user a Twitter user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addTwitterLink(String user) {
         return addLink(R.mipmap.twitter, R.string.twitter, util.openTwitter(user));
     }
@@ -874,6 +943,7 @@ public final class AboutBuilder {
      * @param url any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addGoogleLink(int url) {
         return addGoogleLink(context.getString(url));
     }
@@ -884,6 +954,7 @@ public final class AboutBuilder {
      * @param url any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addGoogleLink(String url) {
         return addLink(R.mipmap.google, R.string.google, url);
     }
@@ -894,6 +965,7 @@ public final class AboutBuilder {
      * @param user a Google Plus user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addGooglePlusLink(int user) {
         return addGooglePlusLink(context.getString(user));
     }
@@ -904,6 +976,7 @@ public final class AboutBuilder {
      * @param user a Google Plus user name or id
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addGooglePlusLink(String user) {
         return addLink(R.mipmap.google_plus, R.string.google_plus, util.openGooglePlus(user));
     }
@@ -914,6 +987,7 @@ public final class AboutBuilder {
      * @param user a Google Play Store user name or id
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addGooglePlayStoreLink(int user) {
         return addGooglePlayStoreLink(context.getString(user));
     }
@@ -924,6 +998,7 @@ public final class AboutBuilder {
      * @param user a Google Play Store user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addGooglePlayStoreLink(String user) {
         return addLink(R.mipmap.google_play_store, R.string.google_play_store, util.openGooglePlayDev(user));
     }
@@ -934,6 +1009,7 @@ public final class AboutBuilder {
      * @param url any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addGoogleGamesLink(int url) {
         return addGoogleGamesLink(context.getString(url));
     }
@@ -944,6 +1020,7 @@ public final class AboutBuilder {
      * @param url any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addGoogleGamesLink(String url) {
         return addLink(R.mipmap.google_play_games, R.string.google_play_games, url);
     }
@@ -954,6 +1031,7 @@ public final class AboutBuilder {
      * @param user a Youtube channel name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addYoutubeChannelLink(int user) {
         return addYoutubeChannelLink(context.getString(user));
     }
@@ -964,10 +1042,10 @@ public final class AboutBuilder {
      * @param user a Youtube channel name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addYoutubeChannelLink(String user) {
         return addLink(R.mipmap.youtube, R.string.youtube, util.openYoutubeChannel(user));
     }
-
 
     /**
      * Adds a Youtube user link on the links section
@@ -975,6 +1053,7 @@ public final class AboutBuilder {
      * @param user a Youtube user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addYoutubeUserLink(int user) {
         return addYoutubeUserLink(context.getString(user));
     }
@@ -985,6 +1064,7 @@ public final class AboutBuilder {
      * @param user a Youtube user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addYoutubeUserLink(String user) {
         return addLink(R.mipmap.youtube, R.string.youtube, util.openYoutubeUser(user));
     }
@@ -995,6 +1075,7 @@ public final class AboutBuilder {
      * @param user a LinkedIn user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLinkedInLink(int user) {
         return addLinkedInLink(context.getString(user));
     }
@@ -1005,6 +1086,7 @@ public final class AboutBuilder {
      * @param user a LinkedIn user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLinkedInLink(String user) {
         return addLink(R.mipmap.linkedin, R.string.linkedin, util.openLinkedIn(user));
     }
@@ -1015,6 +1097,7 @@ public final class AboutBuilder {
      * @param phone a valid skype phone number
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addSkypeLink(int phone) {
         return addSkypeLink(context.getString(phone));
     }
@@ -1025,6 +1108,7 @@ public final class AboutBuilder {
      * @param phone a valid skype phone number
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addSkypeLink(String phone) {
         return addLink(R.mipmap.skype, R.string.skype, util.openSkype(phone));
     }
@@ -1035,6 +1119,7 @@ public final class AboutBuilder {
      * @param phone any phone number
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addWhatsappLink(int name, int phone) {
         return addWhatsappLink(context.getString(name), context.getString(phone));
     }
@@ -1045,6 +1130,7 @@ public final class AboutBuilder {
      * @param phone any phone number
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addWhatsappLink(String name, String phone) {
         return addLink(R.mipmap.whatsapp, R.string.whastapp, util.openAddContact(name, phone));
     }
@@ -1055,6 +1141,7 @@ public final class AboutBuilder {
      * @param url any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAndroidLink(int url) {
         return addAndroidLink(context.getString(url));
     }
@@ -1065,6 +1152,7 @@ public final class AboutBuilder {
      * @param url any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAndroidLink(String url) {
         return addLink(R.mipmap.android, R.string.android, url);
     }
@@ -1075,6 +1163,7 @@ public final class AboutBuilder {
      * @param url any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addDribbbleLink(int url) {
         return addDribbbleLink(context.getString(url));
     }
@@ -1085,6 +1174,7 @@ public final class AboutBuilder {
      * @param url any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addDribbbleLink(String url) {
         return addLink(R.mipmap.dribbble, R.string.dribbble, url);
     }
@@ -1095,6 +1185,7 @@ public final class AboutBuilder {
      * @param url any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addWebsiteLink(int url) {
         return addWebsiteLink(context.getString(url));
     }
@@ -1105,6 +1196,7 @@ public final class AboutBuilder {
      * @param url any url
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addWebsiteLink(String url) {
         return addLink(R.mipmap.website, R.string.website, url);
     }
@@ -1117,6 +1209,7 @@ public final class AboutBuilder {
      * @param message the contect of the e-mail
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addEmailLink(int email, int subject, int message) {
         return addEmailLink(context.getString(email), context.getString(subject), context.getString(message));
     }
@@ -1129,6 +1222,7 @@ public final class AboutBuilder {
      * @param message the contect of the e-mail
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addEmailLink(int email, String subject, String message) {
         return addEmailLink(context.getString(email), subject, message);
     }
@@ -1140,6 +1234,7 @@ public final class AboutBuilder {
      * @param subject the subject of the e-mail
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addEmailLink(int email, String subject) {
         return addEmailLink(context.getString(email), subject, null);
     }
@@ -1151,6 +1246,7 @@ public final class AboutBuilder {
      * @param subject the subject of the e-mail
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addEmailLink(int email, int subject) {
         return addEmailLink(context.getString(email), context.getString(subject), null);
     }
@@ -1163,6 +1259,7 @@ public final class AboutBuilder {
      * @param message the contect of the e-mail
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addEmailLink(String email, String subject, String message) {
         return addLink(R.mipmap.email, R.string.email, util.sendEmail(email, subject, message));
     }
@@ -1173,6 +1270,7 @@ public final class AboutBuilder {
      * @param email addressee e-mail
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addEmailLink(int email) {
         return addEmailLink(context.getString(email));
     }
@@ -1183,10 +1281,10 @@ public final class AboutBuilder {
      * @param email addressee e-mail
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addEmailLink(String email) {
         return addLink(R.mipmap.email, R.string.email, util.sendEmail(email, null, null));
     }
-
 
     /**
      * Adds an action button on the actions section.
@@ -1196,6 +1294,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(Bitmap icon, String label, View.OnClickListener onClickListener) {
         actions.add(new Item(icon, label, onClickListener));
         return this;
@@ -1209,10 +1308,10 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(Bitmap icon, String label, Intent intent) {
         return addAction(icon, label, util.clickIntent(intent));
     }
-
 
     /**
      * Adds an action button on the actions section.
@@ -1222,6 +1321,7 @@ public final class AboutBuilder {
      * @param uri   the action uri
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(Bitmap icon, String label, Uri uri) {
         return addAction(icon, label, util.clickUri(uri));
     }
@@ -1234,7 +1334,7 @@ public final class AboutBuilder {
      * @param url   any url to browse
      * @return the same {@link AboutBuilder} instance
      */
-
+    @NonNull
     public AboutBuilder addAction(Bitmap icon, String label, String url) {
         return addAction(icon, label, Uri.parse(url));
     }
@@ -1247,6 +1347,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(Bitmap icon, int label, View.OnClickListener onClickListener) {
         return addAction(icon, context.getString(label), onClickListener);
     }
@@ -1259,6 +1360,7 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(Bitmap icon, int label, Intent intent) {
         return addAction(icon, label, util.clickIntent(intent));
     }
@@ -1271,6 +1373,7 @@ public final class AboutBuilder {
      * @param uri   the action uri
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(Bitmap icon, int label, Uri uri) {
         return addAction(icon, label, util.clickUri(uri));
     }
@@ -1283,6 +1386,7 @@ public final class AboutBuilder {
      * @param url   any url to browse
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(Bitmap icon, int label, String url) {
         return addAction(icon, label, Uri.parse(url));
     }
@@ -1295,6 +1399,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(int icon, int label, View.OnClickListener onClickListener) {
         return addAction(IconUtil.getBitmap(context, icon), context.getString(label), onClickListener);
     }
@@ -1307,6 +1412,7 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(int icon, int label, Intent intent) {
         return addAction(icon, label, util.clickIntent(intent));
     }
@@ -1319,6 +1425,7 @@ public final class AboutBuilder {
      * @param uri   the action uri
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(int icon, int label, Uri uri) {
         return addAction(icon, label, util.clickUri(uri));
     }
@@ -1331,6 +1438,7 @@ public final class AboutBuilder {
      * @param url   any url to browse
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(int icon, int label, String url) {
         return addAction(icon, label, Uri.parse(url));
     }
@@ -1343,6 +1451,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(int icon, String label, View.OnClickListener onClickListener) {
         return addAction(IconUtil.getBitmap(context, icon), label, onClickListener);
     }
@@ -1355,6 +1464,7 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(int icon, String label, Intent intent) {
         return addAction(icon, label, util.clickIntent(intent));
     }
@@ -1367,6 +1477,7 @@ public final class AboutBuilder {
      * @param uri   the action uri
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(int icon, String label, Uri uri) {
         return addAction(icon, label, util.clickUri(uri));
     }
@@ -1379,6 +1490,7 @@ public final class AboutBuilder {
      * @param url   any url to browse
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addAction(int icon, String label, String url) {
         return addAction(icon, label, Uri.parse(url));
     }
@@ -1391,8 +1503,9 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addAction(Drawable icon, int label, View.OnClickListener onClickListener) {
-        return addAction(IconUtil.getBitmap(context, icon), context.getString(label), onClickListener);
+    @NonNull
+    public AboutBuilder addAction(@NonNull BitmapDrawable icon, int label, View.OnClickListener onClickListener) {
+        return addAction(IconUtil.getBitmap(icon), context.getString(label), onClickListener);
     }
 
     /**
@@ -1403,7 +1516,8 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addAction(Drawable icon, int label, Intent intent) {
+    @NonNull
+    public AboutBuilder addAction(@NonNull BitmapDrawable icon, int label, Intent intent) {
         return addAction(icon, label, util.clickIntent(intent));
     }
 
@@ -1415,7 +1529,8 @@ public final class AboutBuilder {
      * @param uri   the action uri
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addAction(Drawable icon, int label, Uri uri) {
+    @NonNull
+    public AboutBuilder addAction(@NonNull BitmapDrawable icon, int label, Uri uri) {
         return addAction(icon, label, util.clickUri(uri));
     }
 
@@ -1427,7 +1542,8 @@ public final class AboutBuilder {
      * @param url   any url to browse
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addAction(Drawable icon, int label, String url) {
+    @NonNull
+    public AboutBuilder addAction(@NonNull BitmapDrawable icon, int label, String url) {
         return addAction(icon, label, Uri.parse(url));
     }
 
@@ -1439,8 +1555,9 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addAction(Drawable icon, String label, View.OnClickListener onClickListener) {
-        return addAction(IconUtil.getBitmap(context, icon), label, onClickListener);
+    @NonNull
+    public AboutBuilder addAction(@NonNull BitmapDrawable icon, String label, View.OnClickListener onClickListener) {
+        return addAction(IconUtil.getBitmap(icon), label, onClickListener);
     }
 
     /**
@@ -1451,7 +1568,8 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addAction(Drawable icon, String label, Intent intent) {
+    @NonNull
+    public AboutBuilder addAction(@NonNull BitmapDrawable icon, String label, Intent intent) {
         return addAction(icon, label, util.clickIntent(intent));
     }
 
@@ -1463,7 +1581,8 @@ public final class AboutBuilder {
      * @param uri   the action uri
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addAction(Drawable icon, String label, Uri uri) {
+    @NonNull
+    public AboutBuilder addAction(@NonNull BitmapDrawable icon, String label, Uri uri) {
         return addAction(icon, label, util.clickUri(uri));
     }
 
@@ -1475,10 +1594,10 @@ public final class AboutBuilder {
      * @param url   any url to browse
      * @return the same {@link AboutBuilder} instance
      */
-    public AboutBuilder addAction(Drawable icon, String label, String url) {
+    @NonNull
+    public AboutBuilder addAction(@NonNull BitmapDrawable icon, String label, String url) {
         return addAction(icon, label, Uri.parse(url));
     }
-
 
     /**
      * Adds an action button to rate the app on Google Play Store
@@ -1486,6 +1605,7 @@ public final class AboutBuilder {
      * @param appId the application id
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFiveStarsAction(int appId) {
         return addFiveStarsAction(context.getString(appId));
     }
@@ -1496,6 +1616,7 @@ public final class AboutBuilder {
      * @param appId the application id
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFiveStarsAction(String appId) {
         return addAction(R.mipmap.star, R.string.rate_five_stars, util.openPlayStoreAppPage(appId));
     }
@@ -1506,6 +1627,7 @@ public final class AboutBuilder {
      *
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFiveStarsAction() {
         return addFiveStarsAction(getApplicationID());
     }
@@ -1516,6 +1638,7 @@ public final class AboutBuilder {
      * @param appId the application id
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addUpdateAction(int appId) {
         return addUpdateAction(context.getString(appId));
     }
@@ -1527,6 +1650,7 @@ public final class AboutBuilder {
      * @param appId the application id
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addUpdateAction(String appId) {
         return addAction(R.mipmap.update, R.string.update_app, util.openPlayStoreAppPage(appId));
     }
@@ -1537,6 +1661,7 @@ public final class AboutBuilder {
      *
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addUpdateAction() {
         return addUpdateAction(getApplicationID());
     }
@@ -1547,6 +1672,7 @@ public final class AboutBuilder {
      * @param userName the developer user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addMoreFromMeAction(int userName) {
         return addMoreFromMeAction(context.getString(userName));
     }
@@ -1557,6 +1683,7 @@ public final class AboutBuilder {
      * @param userName the developer user name
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addMoreFromMeAction(String userName) {
         return addAction(R.mipmap.google_play_store, R.string.more_apps, util.openPlayStoreAppsList(userName));
     }
@@ -1568,6 +1695,7 @@ public final class AboutBuilder {
      * @param message the content message
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addShareAction(int subject, int message) {
         return addShareAction(context.getString(subject), context.getString(message));
     }
@@ -1579,6 +1707,7 @@ public final class AboutBuilder {
      * @param message the content message
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addShareAction(String subject, String message) {
         return addAction(share, R.string.share_app, util.shareThisApp(subject, message));
     }
@@ -1589,6 +1718,7 @@ public final class AboutBuilder {
      * @param subject the subject
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addShareAction(String subject) {
         return addShareAction(subject, context.getString(R.string.uri_play_store_app_website, context.getPackageName()));
     }
@@ -1599,6 +1729,7 @@ public final class AboutBuilder {
      * @param subject the subject
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addShareAction(int subject) {
         return addShareAction(context.getString(subject));
     }
@@ -1612,6 +1743,7 @@ public final class AboutBuilder {
      * @param content the e-mail content
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFeedbackAction(int email, int subject, int content) {
         return addFeedbackAction(context.getString(email), context.getString(subject), context.getString(content));
     }
@@ -1624,6 +1756,7 @@ public final class AboutBuilder {
      * @param content the e-mail content
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFeedbackAction(int email, String subject, String content) {
         return addAction(R.mipmap.feedback, R.string.feedback_app, util.sendEmail(context.getString(email), subject, content));
     }
@@ -1636,6 +1769,7 @@ public final class AboutBuilder {
      * @param content the e-mail content
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFeedbackAction(String email, String subject, String content) {
         return addAction(R.mipmap.feedback, R.string.feedback_app, util.sendEmail(email, subject, content));
     }
@@ -1647,6 +1781,7 @@ public final class AboutBuilder {
      * @param subject the subject
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFeedbackAction(int email, int subject) {
         return addFeedbackAction(context.getString(email), context.getString(subject));
     }
@@ -1658,10 +1793,10 @@ public final class AboutBuilder {
      * @param subject the subject
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFeedbackAction(String email, String subject) {
         return addFeedbackAction(email, subject, null);
     }
-
 
     /**
      * Adds a feedback action button. No content message will be placed
@@ -1670,10 +1805,10 @@ public final class AboutBuilder {
      * @param subject the subject
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFeedbackAction(int email, String subject) {
         return addFeedbackAction(context.getString(email), subject, null);
     }
-
 
     /**
      * Adds a feedback action button. No subject or content message will be placed
@@ -1681,6 +1816,7 @@ public final class AboutBuilder {
      * @param email the developer e-mail
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFeedbackAction(int email) {
         return addFeedbackAction(context.getString(email));
     }
@@ -1691,6 +1827,7 @@ public final class AboutBuilder {
      * @param email the developer e-mail
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addFeedbackAction(String email) {
         return addFeedbackAction(email, null);
     }
@@ -1701,6 +1838,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addIntroduceAction(View.OnClickListener onClickListener) {
         return addAction(R.mipmap.intrduce, R.string.introduce_app, onClickListener);
     }
@@ -1711,6 +1849,7 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addIntroduceAction(Intent intent) {
         return addIntroduceAction(util.clickIntent(intent));
     }
@@ -1721,6 +1860,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addHelpAction(View.OnClickListener onClickListener) {
         return addAction(R.mipmap.help, R.string.help, onClickListener);
     }
@@ -1731,6 +1871,7 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addHelpAction(Intent intent) {
         return addHelpAction(util.clickIntent(intent));
     }
@@ -1741,6 +1882,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLicenseAction(View.OnClickListener onClickListener) {
         return addAction(R.mipmap.license, R.string.license, onClickListener);
     }
@@ -1751,6 +1893,7 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addLicenseAction(Intent intent) {
         return addLicenseAction(util.clickIntent(intent));
     }
@@ -1761,6 +1904,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addChangeLogAction(View.OnClickListener onClickListener) {
         return addAction(R.mipmap.changelog, R.string.changelog, onClickListener);
     }
@@ -1771,6 +1915,7 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addChangeLogAction(Intent intent) {
         return addChangeLogAction(util.clickIntent(intent));
     }
@@ -1781,6 +1926,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addRemoveAdsAction(View.OnClickListener onClickListener) {
         return addAction(R.mipmap.ads, R.string.remove_ads, onClickListener);
     }
@@ -1792,10 +1938,10 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addRemoveAdsAction(Intent intent) {
         return addRemoveAdsAction(util.clickIntent(intent));
     }
-
 
     /**
      * Adds an donate action button
@@ -1803,6 +1949,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addDonateAction(View.OnClickListener onClickListener) {
         return addAction(R.mipmap.donate, R.string.donate, onClickListener);
     }
@@ -1813,17 +1960,19 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addDonateAction(Intent intent) {
         return addDonateAction(util.clickIntent(intent));
     }
 
     /**
-     * Wraps the content in a {@link android.support.v7.widget.CardView}
+     * Wraps the content in a {@link CardView}
      *
      * @param showAsCard true if show in a CardView
      * @return the same {@link AboutBuilder} instance
      */
 
+    @NonNull
     public AboutBuilder setShowAsCard(boolean showAsCard) {
         this.showAsCard = showAsCard;
         return this;
@@ -1835,6 +1984,7 @@ public final class AboutBuilder {
      * @param url the url to privacy policy web page
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addPrivacyPolicyAction(String url) {
         return addAction(R.mipmap.privacy, R.string.privacy, util.intent(url));
     }
@@ -1845,6 +1995,7 @@ public final class AboutBuilder {
      * @param onClickListener the click callback
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addPrivacyPolicyAction(View.OnClickListener onClickListener) {
         return addAction(R.mipmap.privacy, R.string.privacy, onClickListener);
     }
@@ -1855,6 +2006,7 @@ public final class AboutBuilder {
      * @param intent the action intent
      * @return the same {@link AboutBuilder} instance
      */
+    @NonNull
     public AboutBuilder addPrivacyPolicyAction(Intent intent) {
         return addAction(R.mipmap.privacy, R.string.privacy, util.clickIntent(intent));
     }
@@ -1951,14 +2103,17 @@ public final class AboutBuilder {
         return wrapScrollView;
     }
 
+    @NonNull
     public LinkedList<Item> getLinks() {
         return links;
     }
 
+    @NonNull
     public LinkedList<Item> getActions() {
         return actions;
     }
 
+    @NonNull
     public AboutView build() {
         AboutView aboutView = new AboutView(context);
         aboutView.build(this);
